+++ lib/td_dd/data_structures/data_structures.ex
@@ -10,6 +10,7 @@ defmodule TdDd.DataStructures do
   alias TdCache.StructureCache
   alias TdCache.TemplateCache
   alias TdDd.DataStructures.DataStructure
+  alias TdDd.DataStructures.DataStructureRelation
   alias TdDd.DataStructures.DataStructureVersion
   alias TdDd.Repo
   alias TdDd.Utils.CollectionUtils
@@ -26,13 +27,14 @@ defmodule TdDd.DataStructures do
       [%DataStructure{}, ...]
 
   """
-  def list_data_structures(params \\ %{}) do
+  def list_data_structures(params \\ %{}, options \\ []) do
     filter = build_filter(DataStructure, params)
 
     DataStructure
     |> preload([ds], [:system])
     |> where([ds], ^filter)
     |> Repo.all()
+    |> Enum.map(&enrich(&1, options))
   end
 
   def list_data_structures_with_no_parents(params \\ %{}, options \\ []) do
@@ -40,8 +42,11 @@ defmodule TdDd.DataStructures do
 
     DataStructure
     |> where([ds], ^filter)
-    |> where([ds], is_nil(ds.class) or ds.class != "field")
-    |> with_deleted(Keyword.get(options, :deleted, true))
+    |> join(:left, [ds], dsv in assoc(ds, :versions))
+    |> where([ds, dsv], is_nil(dsv.class) or dsv.class != "field")
+    |> with_deleted(options, dynamic([_, dsv], is_nil(dsv.deleted_at)))
+    |> select([ds, dsv], ds)
+    |> distinct(true)
     |> Repo.all()
     |> Repo.preload(system: [], versions: :parents)
     |> Enum.filter(&latest_version_is_root?/1)
@@ -105,112 +110,109 @@ defmodule TdDd.DataStructures do
     |> Repo.preload(:system)
   end
 
-  def get_data_structure_version!(data_structure_id, version) do
+  def get_data_structure_version!(data_structure_id, version, options) do
     attrs = %{data_structure_id: data_structure_id, version: version}
 
     DataStructureVersion
     |> Repo.get_by!(attrs)
     |> Repo.preload(data_structure: :system)
+    |> enrich(options)
   end
 
-  def get_data_structure_version!(data_structure_version_id) do
+  def get_data_structure_version!(data_structure_version_id, options) do
     DataStructureVersion
     |> Repo.get!(data_structure_version_id)
     |> Repo.preload(data_structure: :system)
+    |> enrich(options)
   end
 
-  def get_data_structure_with_fields!(data_structure_id, options \\ []) do
-    data_structure_id
-    |> get_data_structure!
-    |> with_latest_fields(options)
+  defp enrich(nil, _), do: nil
+
+  defp enrich(%DataStructure{} = ds, options) do
+    ds
+    |> enrich(options, :versions, fn ds -> Repo.preload(ds, :versions) end)
+    |> enrich(options, :latest, fn ds -> get_latest_version(ds) end)
   end
 
-  def get_latest_fields(data_structure_id, options \\ []) do
-    data_structure_id
-    |> get_latest_version
-    |> get_field_structures(options)
+  defp enrich(%DataStructureVersion{} = dsv, options) do
+    dsv
+    |> enrich(options, :parents, fn dsv -> get_parents(dsv, deleted: false) end)
+    |> enrich(options, :children, fn dsv -> get_children(dsv, deleted: false) end)
+    |> enrich(options, :siblings, fn dsv -> get_siblings(dsv, deleted: false) end)
+    |> enrich(options, :data_fields, fn dsv -> get_field_structures(dsv, deleted: false) end)
+    |> enrich(options, :data_field_external_ids, fn dsv -> get_field_external_ids(dsv) end)
+    |> enrich(options, :data_field_links, fn dsv -> get_field_links(dsv) end)
+    |> enrich(options, :versions, fn dsv -> get_versions(dsv) end)
+    |> enrich(options, :system, fn dsv ->
+      dsv |> Repo.preload(data_structure: :system) |> Map.get(:data_structure) |> Map.get(:system)
+    end)
+    |> enrich(options, :ancestry, fn dsv -> get_ancestry(dsv) end)
+    |> enrich(options, :path, fn dsv -> get_path(dsv) end)
+    |> enrich(options, :links, fn %{data_structure_id: id} -> get_structure_links(id) end)
   end
+
+  defp enrich(%{} = target, options, key, fun) do
+    target_key = get_target_key(key)
+
+    case Enum.member?(options, key) do
+      false -> target
+      true -> Map.put(target, target_key, fun.(target))
+    end
+  end
+
+  defp get_target_key(:data_field_external_ids), do: :data_fields
+  defp get_target_key(:data_field_links), do: :data_fields
+  defp get_target_key(key), do: key
 
   def get_field_structures(data_structure_version, options \\ []) do
     data_structure_version
     |> Ecto.assoc(:children)
-    |> join(:inner, [child], ds in assoc(child, :data_structure))
-    |> with_deleted(Keyword.get(options, :deleted, true), 4)
-    |> where([_child, _parent, _rel, ds], ds.class == "field")
-    |> select([_child, _parent, _rel, ds], ds)
+    |> where([child, _parent, _rel], child.class == "field")
+    |> with_deleted(options, dynamic([child, _parent, _rel], is_nil(child.deleted_at)))
+    # |> join(:inner, [child, _parent, _rel], ds in assoc(child, :data_structure))
+    # |> select([_child, _parent, _rel, ds], ds)
+    |> select([child, _parent, _rel], child)
     |> Repo.all()
   end
 
-  def get_latest_children(data_structure_id, options \\ []) do
-    data_structure_id
-    |> get_latest_version
-    |> get_children(options)
-  end
-
-  def get_children(data_structure_version, options \\ []) do
-    data_structure_version
-    |> Ecto.assoc([:children, :data_structure])
-    |> with_deleted(Keyword.get(options, :deleted, true))
+  def get_children(%DataStructureVersion{id: id}, options \\ []) do
+    DataStructureRelation
+    |> where([r], r.parent_id == ^id)
+    |> join(:inner, [r], parent in assoc(r, :child))
+    |> with_deleted(options, dynamic([_, parent], is_nil(parent.deleted_at)))
+    # |> join(:inner, [_, parent], ds in assoc(parent, :data_structure))
+    # |> select([_, _, ds], ds)
+    |> select([_, child], child)
+    |> distinct(true)
     |> Repo.all()
-    |> Repo.preload(:system)
+    |> Repo.preload(data_structure: :system)
   end
 
-  defp with_deleted(query, true), do: query
-
-  defp with_deleted(query, _deleted) do
-    where(query, [ds], is_nil(ds.deleted_at))
-  end
-
-  defp with_deleted(query, true, _), do: query
-
-  defp with_deleted(query, _deleted, 4) do
-    where(query, [_, _, _, ds], is_nil(ds.deleted_at))
-  end
-
-  def get_latest_parents(data_structure_id, options \\ []) do
-    data_structure_id
-    |> get_latest_version
-    |> get_parents(options)
-  end
-
-  def get_parents(data_structure_version, options \\ []) do
-    data_structure_version
-    |> Ecto.assoc([:parents, :data_structure])
-    |> with_deleted(Keyword.get(options, :deleted, true))
+  def get_parents(%DataStructureVersion{id: id}, options \\ []) do
+    DataStructureRelation
+    |> where([r], r.child_id == ^id)
+    |> join(:inner, [r], parent in assoc(r, :parent))
+    |> with_deleted(options, dynamic([_, parent], is_nil(parent.deleted_at)))
+    # |> join(:inner, [_, parent], ds in assoc(parent, :data_structure))
+    # |> select([_, _, ds], ds)
+    |> select([_, parent], parent)
+    |> distinct(true)
     |> Repo.all()
-    |> Repo.preload(:system)
+    |> Repo.preload(data_structure: :system)
   end
 
-  defp get_latest_siblings(data_structure_id, options) do
-    data_structure_id
-    |> get_latest_version
-    |> get_siblings(options)
-  end
-
-  def get_siblings(data_structure_version, options \\ []) do
-    data_structure_version
-    |> Ecto.assoc(:parents)
+  def get_siblings(%DataStructureVersion{id: id}, options \\ []) do
+    DataStructureRelation
+    |> where([r], r.child_id == ^id)
+    |> join(:inner, [r], parent in assoc(r, :parent))
+    |> join(:inner, [_, parent], child in assoc(parent, :children))
+    |> with_deleted(options, dynamic([_, parent, _], is_nil(parent.deleted_at)))
+    |> with_deleted(options, dynamic([_, _, child], is_nil(child.deleted_at)))
+    # |> join(:inner, [_, _, child], ds in assoc(child, :data_structure))
+    # |> select([_, _, _, ds], ds)
+    |> select([_, _, child], child)
+    |> distinct(true)
     |> Repo.all()
-    |> Enum.filter(&is_active_data_structure_version(&1, Keyword.get(options, :deleted, true)))
-    |> Enum.flat_map(&get_children(&1, options))
-    |> Enum.uniq()
-  end
-
-  defp is_active_data_structure_version(data_structure_version, false) do
-    data_structure =
-      data_structure_version
-      |> Repo.preload(:data_structure)
-      |> Map.get(:data_structure)
-
-    is_nil(data_structure.deleted_at)
-  end
-
-  defp is_active_data_structure_version(_data_structure_version, _true), do: true
-
-  def get_latest_ancestry(data_structure_id) do
-    data_structure_id
-    |> get_latest_version
-    |> get_ancestry
   end
 
   def get_versions(%DataStructureVersion{} = dsv) do
@@ -222,41 +224,6 @@ defmodule TdDd.DataStructures do
   def with_versions(%DataStructure{} = data_structure) do
     data_structure
     |> Repo.preload(:versions)
-  end
-
-  def with_latest_fields(%{id: id} = data_structure, options \\ []) do
-    fields = get_latest_fields(id, options)
-
-    data_structure
-    |> Map.put(:data_fields, fields)
-  end
-
-  def with_latest_children(%{id: id} = data_structure, options \\ []) do
-    children = get_latest_children(id, options)
-
-    data_structure
-    |> Map.put(:children, children)
-  end
-
-  def with_latest_parents(%{id: id} = data_structure, options \\ []) do
-    parents = get_latest_parents(id, options)
-
-    data_structure
-    |> Map.put(:parents, parents)
-  end
-
-  def with_latest_siblings(%{id: id} = data_structure, options \\ []) do
-    siblings = get_latest_siblings(id, options)
-
-    data_structure
-    |> Map.put(:siblings, siblings)
-  end
-
-  def with_latest_ancestry(%{id: id} = data_structure) do
-    ancestry = get_latest_ancestry(id)
-
-    data_structure
-    |> Map.put(:ancestry, ancestry)
   end
 
   @doc """
@@ -272,25 +239,31 @@ defmodule TdDd.DataStructures do
 
   """
   def create_data_structure(attrs \\ %{}) do
-    result =
-      %DataStructure{}
-      |> DataStructure.changeset(attrs)
-      |> Repo.insert()
+    {dsv_attrs, ds_attrs} =
+      Map.split(attrs, ["class", "description", "metadata", "group", "name", "type", "version"])
 
-    case result do
-      {:ok, data_structure} ->
-        %DataStructureVersion{data_structure_id: data_structure.id, version: 0}
-        |> Repo.insert()
+    Repo.transaction(fn ->
+      with {:ok, data_structure} <- insert_data_structure(ds_attrs),
+           {:ok, _dsv} <- insert_data_structure_version(data_structure, dsv_attrs) do
+        @search_service.put_search(data_structure)
+        Repo.preload(data_structure, :system)
+      else
+        {:error, e} -> Repo.rollback(e)
+        e -> Repo.rollback(e)
+      end
+    end)
+  end
 
-        data_structure
-        |> with_latest_fields
-        |> @search_service.put_search
+  defp insert_data_structure(attrs) do
+    %DataStructure{}
+    |> DataStructure.changeset(attrs)
+    |> Repo.insert()
+  end
 
-        {:ok, data_structure |> Repo.preload(:system)}
-
-      _ ->
-        result
-    end
+  defp insert_data_structure_version(%DataStructure{id: data_structure_id}, attrs) do
+    %DataStructureVersion{data_structure_id: data_structure_id, version: 0}
+    |> DataStructureVersion.changeset(attrs)
+    |> Repo.insert()
   end
 
   @doc """
@@ -305,11 +278,10 @@ defmodule TdDd.DataStructures do
       {:error, %Ecto.Changeset{}}
 
   """
-  def update_data_structure(
-        %DataStructure{type: type} = data_structure,
-        %{"df_content" => content} = attrs
-      )
+  def update_data_structure(%DataStructure{} = data_structure, %{"df_content" => content} = attrs)
       when not is_nil(content) do
+    %{type: type} = get_latest_version(data_structure)
+
     case TemplateCache.get_by_name!(type) do
       %{:content => content_schema} ->
         content_changeset = Validation.build_changeset(content, content_schema)
@@ -337,7 +309,6 @@ defmodule TdDd.DataStructures do
     case result do
       {:ok, data_structure} ->
         data_structure
-        |> with_latest_fields
         |> @search_service.put_search
 
         result
@@ -385,18 +356,22 @@ defmodule TdDd.DataStructures do
     DataStructure.changeset(data_structure, %{})
   end
 
+  def get_latest_version(target, options \\ [])
+
   @doc """
   Returns the latest data structure version for a given data structure.
   """
-  def get_latest_version(%DataStructure{versions: versions}) when is_list(versions) do
-    Enum.max_by(versions, & &1.version)
+  def get_latest_version(%DataStructure{versions: versions}, options) when is_list(versions) do
+    versions
+    |> Enum.max_by(& &1.version)
+    |> enrich(options)
   end
 
   @doc """
   Returns the latest data structure version for a given data structure.
   """
-  def get_latest_version(%DataStructure{id: id}) do
-    get_latest_version(id)
+  def get_latest_version(%DataStructure{id: id}, options) do
+    get_latest_version(id, options)
   end
 
   @doc """
@@ -408,15 +383,17 @@ defmodule TdDd.DataStructures do
       %DataStructureVersion{}
 
   """
-  def get_latest_version(data_structure_id) do
+  def get_latest_version(data_structure_id, options) do
     Repo.one(
       from(dsv in DataStructureVersion,
         where: dsv.data_structure_id == type(^data_structure_id, :integer),
         order_by: [desc: :version],
         limit: 1,
-        select: dsv
+        select: dsv,
+        preload: :data_structure
       )
     )
+    |> enrich(options)
   end
 
   def add_domain_id(%{"ou" => domain_name, "domain_id" => nil} = data, domain_map) do
@@ -439,20 +416,28 @@ defmodule TdDd.DataStructures do
     Repo.get_by(DataStructure, clauses)
   end
 
-  def with_field_external_ids(%{data_fields: data_fields} = data_structure) do
-    data_structure
-    |> Map.put(
-      :data_fields,
-      Enum.map(
-        data_fields,
-        &Map.put(
-          &1,
-          :external_id,
-          StructureCache.get_external_id(
-            data_structure.system.external_id,
-            &1.external_id
-          )
+  def get_field_external_ids(%{data_fields: data_fields}) do
+    data_fields
+    |> Repo.preload(data_structure: :system)
+    |> Enum.map(
+      &Map.put(
+        &1,
+        :external_id,
+        StructureCache.get_external_id(
+          &1.data_structure.system.external_id,
+          &1.data_structure.external_id
         )
+      )
+    )
+  end
+
+  def get_field_links(%{data_fields: data_fields}) do
+    data_fields
+    |> Enum.map(
+      &Map.put(
+        &1,
+        :links,
+        get_structure_links(&1.data_structure_id)
       )
     )
   end
@@ -464,19 +449,10 @@ defmodule TdDd.DataStructures do
     end
   end
 
-  def with_latest_path(%DataStructure{} = data_structure) do
-    path =
-      data_structure
-      |> get_latest_path
-
-    data_structure
-    |> Map.put(:path, path)
-  end
-
   def get_latest_path(%DataStructure{id: id}) do
     id
-    |> get_latest_version
-    |> get_path
+    |> get_latest_version()
+    |> get_path()
   end
 
   def get_path(%DataStructureVersion{} = dsv) do
@@ -486,30 +462,29 @@ defmodule TdDd.DataStructures do
     |> Enum.reverse()
   end
 
-  def get_ancestry(%DataStructureVersion{parents: %NotLoaded{}} = data_structure_version) do
+  defp get_ancestry(%DataStructureVersion{parents: %NotLoaded{}} = data_structure_version) do
     data_structure_version
-    |> Repo.preload([:parents, :data_structure])
+    |> Repo.preload(:parents)
     |> get_ancestry
   end
 
-  def get_ancestry(%DataStructureVersion{parents: [], data_structure: data_structure}) do
-    [data_structure]
+  defp get_ancestry(%DataStructureVersion{parents: []}) do
+    []
   end
 
-  def get_ancestry(%DataStructureVersion{parents: parents, data_structure: data_structure}) do
+  defp get_ancestry(%DataStructureVersion{parents: parents}) do
     parent =
       case get_first_active_parent(parents) do
         nil -> hd(parents)
         parent -> parent
       end
 
-    [data_structure | get_ancestry(parent)]
+    [parent | get_ancestry(parent)]
   end
 
   defp get_first_active_parent(parents) do
     parents
-    |> Repo.preload(:data_structure)
-    |> Enum.find(&(&1.data_structure.deleted_at == nil))
+    |> Enum.find(&(&1.deleted_at == nil))
   end
 
   def get_structure_by_external_ids(system_external_id, external_id) do
@@ -518,5 +493,26 @@ defmodule TdDd.DataStructures do
     |> where([_, s], s.external_id == ^system_external_id)
     |> where([d, _], d.external_id == ^external_id)
     |> Repo.one()
+  end
+
+  def get_structure_version_by_hash_and_external_id(hash, external_id) do
+    DataStructureVersion
+    |> join(:inner, [data_structure], ds in assoc(data_structure, :data_structure))
+    |> where([_, ds], ds.external_id == ^external_id)
+    |> where([dsv, _], dsv.hash == ^hash)
+    |> preload([:data_structure])
+    |> Repo.one()
+  end
+
+  defp with_deleted(query, options, dynamic) when is_list(options) do
+    include_deleted = Keyword.get(options, :deleted, true)
+    with_deleted(query, include_deleted, dynamic)
+  end
+
+  defp with_deleted(query, true, _), do: query
+
+  defp with_deleted(query, _false, dynamic) do
+    query
+    |> where(^dynamic)
   end
 end